+++ fuc/api/common.py
@@ -23,6 +23,391 @@ import matplotlib.patches as mpatches
 
 FUC_PATH = pathlib.Path(__file__).parent.parent.parent.absolute()
 
+class AnnFrame:
+    """
+    Class for storing sample annotation data.
+
+    This class stores sample annotation data as :class:`pandas.DataFrame`
+    with sample names as index.
+
+    Note that an AnnFrame can have a different set of samples than its
+    accompanying MafFrame.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        DataFrame containing sample annotation data. The index must be
+        sample names.
+
+    See Also
+    --------
+    AnnFrame.from_dict
+        Construct AnnFrame from dict of array-like or dicts.
+    AnnFrame.from_file
+        Construct AnnFrame from a delimited text file.
+
+    Examples
+    --------
+
+    >>> import pandas as pd
+    >>> from fuc import pymaf
+    >>> data = {
+    ...     'Tumor_Sample_Barcode': ['Steven_N', 'Steven_T', 'Sara_N', 'Sara_T'],
+    ...     'Subject': ['Steven', 'Steven', 'Sara', 'Sara'],
+    ...     'Type': ['Normal', 'Tumor', 'Normal', 'Tumor'],
+    ...     'Age': [30, 30, 57, 57]
+    ... }
+    >>> df = pd.DataFrame(data)
+    >>> df = df.set_index('Tumor_Sample_Barcode')
+    >>> af = pymaf.AnnFrame(df)
+    >>> af.df
+                         Subject    Type  Age
+    Tumor_Sample_Barcode
+    Steven_N              Steven  Normal   30
+    Steven_T              Steven   Tumor   30
+    Sara_N                  Sara  Normal   57
+    Sara_T                  Sara   Tumor   57
+    """
+
+    def _check_df(self, df):
+        if type(df.index) == pd.RangeIndex:
+            m = "Index must be sample names, not 'pandas.RangeIndex'."
+            raise ValueError(m)
+        if df.isin([np.inf]).any().any():
+            raise ValueError('Found positive infinity.')
+        if df.isin([-np.inf]).any().any():
+            raise ValueError('Found negative infinity.')
+        return df
+
+    def __init__(self, df):
+        self._df = self._check_df(df)
+
+    @property
+    def df(self):
+        """pandas.DataFrame : DataFrame containing sample annotation data."""
+        return self._df
+
+    @df.setter
+    def df(self, value):
+        self._df = self._check_df(value)
+
+    @property
+    def samples(self):
+        """list : List of the sample names."""
+        return list(self.df.index.to_list())
+
+    @property
+    def shape(self):
+        """tuple : Dimensionality of AnnFrame (samples, annotations)."""
+        return self.df.shape
+
+    def filter_mf(self, mf):
+        """
+        Filter the AnnFrame for the samples in the MafFrame.
+
+        Parameters
+        ----------
+        mf : MafFrame
+            MafFrame containing target samples.
+
+        Returns
+        -------
+        AnnFrame
+            Filtered AnnFrame object.
+        """
+        df = self.df.loc[mf.samples]
+        return self.__class__(df)
+
+    @classmethod
+    def from_dict(cls, data, sample_col='Tumor_Sample_Barcode'):
+        """Construct AnnFrame from dict of array-like or dicts.
+
+        The dictionary must have at least one column that represents sample
+        names which are used as index for pandas.DataFrame.
+
+        Parameters
+        ----------
+        data : dict
+            Of the form {field : array-like} or {field : dict}.
+        sample_col : str, default: 'Tumor_Sample_Barcode'
+            Column containing sample names.
+
+        Returns
+        -------
+        AnnFrame
+            AnnFrame object.
+
+        See Also
+        --------
+        AnnFrame
+            AnnFrame object creation using constructor.
+        AnnFrame.from_file
+            Construct AnnFrame from a delimited text file.
+
+        Examples
+        --------
+
+        >>> from fuc import pymaf
+        >>> data = {
+        ...     'Tumor_Sample_Barcode': ['Steven_Normal', 'Steven_Tumor', 'Sara_Normal', 'Sara_Tumor'],
+        ...     'Subject': ['Steven', 'Steven', 'Sara', 'Sara'],
+        ...     'Type': ['Normal', 'Tumor', 'Normal', 'Tumor'],
+        ...     'Age': [30, 30, 57, 57]
+        ... }
+        >>> af = pymaf.AnnFrame.from_dict(data)
+        >>> af.df
+                             Subject    Type  Age
+        Tumor_Sample_Barcode
+        Steven_Normal         Steven  Normal   30
+        Steven_Tumor          Steven   Tumor   30
+        Sara_Normal             Sara  Normal   57
+        Sara_Tumor              Sara   Tumor   57
+        """
+        df = pd.DataFrame(data)
+        df = df.set_index(sample_col)
+        return cls(df)
+
+    @classmethod
+    def from_file(cls, fn, sample_col='Tumor_Sample_Barcode', sep='\t'):
+        """
+        Construct AnnFrame from a delimited text file.
+
+        The text file must have at least one column that represents
+        sample names which are used as index for pandas.DataFrame.
+
+        Parameters
+        ----------
+        fn : str
+            Text file path (zipped or unzipped).
+        sample_col : str, default: 'Tumor_Sample_Barcode'
+            Column containing sample names.
+        sep : str, default: '\\\\t'
+            Delimiter to use.
+
+        Returns
+        -------
+        AnnFrame
+            AnnFrame.
+
+        See Also
+        --------
+        AnnFrame
+            AnnFrame object creation using constructor.
+        AnnFrame.from_dict
+            Construct AnnFrame from dict of array-like or dicts.
+
+        Examples
+        --------
+
+        >>> from fuc import pymaf
+        >>> af1 = pymaf.AnnFrame.from_file('sample-annot-1.tsv')
+        >>> af2 = pymaf.AnnFrame.from_file('sample-annot-2.csv', sample_col='SampleID', sep=',')
+        """
+        df = pd.read_table(fn, sep=sep)
+        df = df.set_index(sample_col)
+        return cls(df)
+
+    def legend_handles(
+        self, col, samples=None, numeric=False, segments=5, decimals=0,
+        cmap='Pastel1'
+    ):
+        """Return legend handles for the given column.
+
+        In the case of a numeric column, use ``numeric=True`` which will
+        divide the values into equal-sized intervals, with the number of
+        intervals determined by the `segments` option.
+
+        Parameters
+        ----------
+        col : str
+            Column name.
+        samples : list, optional
+            If provided, these samples will be used to create legend handles.
+        numeric : bool, default: False
+            If True, the column will be treated as numeric.
+        segments : int, default: 5
+            If ``numeric`` is True, the numbers will be divided
+            into this number of equal-sized segments.
+        decimals : int, default: 0
+            If ``numeric`` is True, the numbers will be rounded up to this
+            number of decimals.
+        cmap : str, default: 'Pastel1'
+            Color map.
+
+        Returns
+        -------
+        list
+            Legend handles.
+
+        See Also
+        --------
+        AnnFrame.plot_annot
+            Create a 1D categorical heatmap for the given column.
+
+        Examples
+        --------
+
+        .. plot::
+
+            >>> import matplotlib.pyplot as plt
+            >>> from fuc import common, pymaf
+            >>> common.load_dataset('tcga-laml')
+            >>> f = '~/fuc-data/tcga-laml/tcga_laml_annot.tsv'
+            >>> af = pymaf.AnnFrame.from_file(f)
+            >>> fig, ax = plt.subplots()
+            >>> handles1 = af.legend_handles('FAB_classification',
+            ...                              cmap='Dark2')
+            >>> handles2 = af.legend_handles('days_to_last_followup',
+            ...                              numeric=True,
+            ...                              cmap='viridis')
+            >>> handles3 = af.legend_handles('Overall_Survival_Status')
+            >>> leg1 = ax.legend(handles=handles1,
+            ...                  title='FAB_classification',
+            ...                  loc='center left')
+            >>> leg2 = ax.legend(handles=handles2,
+            ...                  title='days_to_last_followup',
+            ...                  loc='center')
+            >>> leg3 = ax.legend(handles=handles3,
+            ...                  title='Overall_Survival_Status',
+            ...                  loc='center right')
+            >>> ax.add_artist(leg1)
+            >>> ax.add_artist(leg2)
+            >>> ax.add_artist(leg3)
+            >>> plt.tight_layout()
+        """
+        s, l = self._get_col(col, numeric=numeric, segments=segments,
+            samples=samples)
+        colors = plt.cm.get_cmap(cmap)(np.linspace(0, 1, len(l)))
+        handles = []
+        for i, label in enumerate(l):
+            handles.append(mpatches.Patch(color=colors[i], label=label))
+        return handles
+
+    def _get_col(
+        self, col, numeric=False, segments=5, samples=None, decimals=0
+    ):
+        s = self.df[col]
+        if numeric:
+            boundaries = list(np.linspace(s.min(), s.max(),
+                segments+1, endpoint=True))
+            intervals = list(zip(boundaries[:-1], boundaries[1:]))
+            def f(x):
+                if pd.isna(x):
+                    return x
+                for i, interval in enumerate(intervals):
+                    a, b = interval
+                    if a <= x <= b:
+                        return f'G{i} ({b:.{decimals}f})'
+            s = s.apply(f)
+        if samples is not None:
+            s = s[samples]
+        l = sorted([x for x in s.unique() if x == x])
+        return s, l
+
+    def plot_annot(
+        self, col, samples=None, numeric=False, segments=5, decimals=0,
+        cmap='Pastel1', ax=None, figsize=None, **kwargs
+    ):
+        """
+        Create a 1D categorical heatmap of the column.
+
+        In the case of a numeric column, set ``numeric`` as True which will
+        divide the values into equal-sized intervals, with the number of
+        intervals determined by ``segments``.
+
+        See the :ref:`tutorials:Create customized oncoplots` tutorial to
+        learn how to create customized oncoplots.
+
+        Parameters
+        ----------
+        col : str
+            Column name.
+        samples : list, optional
+            If provided, these samples will be used to create legend handles.
+        numeric : bool, default: False
+            If True, the column will be treated as numeric.
+        segments : int, default: 5
+            If ``numeric`` is True, the numbers will be divided
+            into this number of equal-sized segments.
+        decimals : int, default: 0
+            If ``numeric`` is True, the numbers will be rounded up to this
+            number of decimals.
+        cmap : str, default: 'Pastel1'
+            Color map.
+        ax : matplotlib.axes.Axes, optional
+            Pre-existing axes for the plot. Otherwise, crete a new one.
+        figsize : tuple, optional
+            Width, height in inches. Format: (float, float).
+
+        Returns
+        -------
+        list
+            Legend handles.
+
+        See Also
+        --------
+        AnnFrame.legend_handles
+            Return legend handles for the given column.
+
+        Examples
+        --------
+
+        .. plot::
+
+            >>> import matplotlib.pyplot as plt
+            >>> from fuc import common, pymaf
+            >>> common.load_dataset('tcga-laml')
+            >>> f = '~/fuc-data/tcga-laml/tcga_laml_annot.tsv'
+            >>> af = pymaf.AnnFrame.from_file(f)
+            >>> fig, [ax1, ax2, ax3] = plt.subplots(3, 1, figsize=(10, 5))
+            >>> af.plot_annot('FAB_classification', ax=ax1, linewidths=1, cmap='Dark2')
+            >>> af.plot_annot('days_to_last_followup', ax=ax2, linewidths=1, cmap='viridis')
+            >>> af.plot_annot('Overall_Survival_Status', ax=ax3, linewidths=1)
+            >>> ax1.set_xlabel('')
+            >>> ax2.set_xlabel('')
+            >>> ax1.set_ylabel('')
+            >>> ax2.set_ylabel('')
+            >>> ax3.set_ylabel('')
+            >>> plt.tight_layout()
+            >>> plt.subplots_adjust(wspace=0.01, hspace=0.01)
+        """
+        s, l = self._get_col(col, numeric=numeric, segments=segments,
+            samples=samples)
+        d = {k: v for v, k in enumerate(l)}
+        df = s.to_frame().applymap(lambda x: x if pd.isna(x) else d[x])
+
+        # Determine which matplotlib axes to plot on.
+        if ax is None:
+            fig, ax = plt.subplots(figsize=figsize)
+
+        sns.heatmap(df.T, ax=ax, cmap=cmap, cbar=False, **kwargs)
+
+        ax.set_xlabel('Samples')
+        ax.set_ylabel(col)
+        ax.set_xticks([])
+        ax.set_yticks([])
+
+        return ax
+
+    def sorted_samples(self, by, mf=None, keep_empty=False, nonsyn=False):
+        """
+        Return a sorted list of sample names.
+
+        Parameters
+        ----------
+        df : str or list
+            Column or list of columns to sort by.
+        """
+        df = self.df.copy()
+
+        if nonsyn:
+            samples = mf.matrix_waterfall(keep_empty=keep_empty).columns
+            df = df.loc[samples]
+
+        df = df.sort_values(by=by)
+
+        return df.index.to_list()
+
 def _script_name():
     """Return the current script's filename."""
     fn = inspect.stack()[1].filename